+++ tests/rule_based_profiler/conftest.py
@@ -1,7 +1,7 @@
 import datetime
 import os
 import shutil
-from typing import Any, Dict, List
+from typing import Any, Dict, List, Optional
 
 import numpy as np
 import pandas as pd
@@ -24,6 +24,140 @@ from great_expectations.rule_based_profiler.parameter_builder import (
 from great_expectations.rule_based_profiler.rule import Rule
 
 yaml = YAML()
+
+
+# TODO: AJB 20210525 This fixture is not yet used but may be helpful to generate batches for unit tests of multibatch
+#  workflows.  It should probably be extended to add different column types / data.
+@pytest.fixture
+def multibatch_generic_csv_generator():
+    """
+    Construct a series of csv files with many data types for use in multibatch testing
+    """
+
+    def _multibatch_generic_csv_generator(
+        data_path: str,
+        start_date: Optional[datetime.datetime] = None,
+        num_event_batches: Optional[int] = 20,
+        num_events_per_batch: Optional[int] = 5,
+    ) -> List[str]:
+
+        if start_date is None:
+            start_date = datetime.datetime(2000, 1, 1)
+
+        file_list = []
+        category_strings = {
+            0: "category0",
+            1: "category1",
+            2: "category2",
+            3: "category3",
+            4: "category4",
+            5: "category5",
+            6: "category6",
+        }
+        for batch_num in range(num_event_batches):
+            # generate a dataframe with multiple column types
+            batch_start_date = start_date + datetime.timedelta(
+                days=(batch_num * num_events_per_batch)
+            )
+            # TODO: AJB 20210416 Add more column types
+            df = pd.DataFrame(
+                {
+                    "event_date": [
+                        (batch_start_date + datetime.timedelta(days=i)).strftime(
+                            "%Y-%m-%d"
+                        )
+                        for i in range(num_events_per_batch)
+                    ],
+                    "batch_num": [batch_num + 1 for _ in range(num_events_per_batch)],
+                    "string_cardinality_3": [
+                        category_strings[i % 3] for i in range(num_events_per_batch)
+                    ],
+                }
+            )
+            filename = f"csv_batch_{batch_num + 1:03}_of_{num_event_batches:03}.csv"
+            file_list.append(filename)
+            # noinspection PyTypeChecker
+            df.to_csv(
+                os.path.join(data_path, filename),
+                index_label="intra_batch_index",
+            )
+
+        return file_list
+
+    return _multibatch_generic_csv_generator
+
+
+@pytest.fixture
+def multibatch_generic_csv_generator_context(monkeypatch, empty_data_context):
+    context: DataContext = empty_data_context
+    monkeypatch.chdir(context.root_directory)
+    data_relative_path = "../data"
+    data_path = os.path.join(context.root_directory, data_relative_path)
+    os.makedirs(data_path, exist_ok=True)
+
+    data_connector_base_directory = "./"
+    monkeypatch.setenv("base_directory", data_connector_base_directory)
+    monkeypatch.setenv("data_fixtures_root", data_relative_path)
+
+    datasource_name = "generic_csv_generator"
+    data_connector_name = "daily_data_connector"
+    asset_name = "daily_data_asset"
+    datasource_config = fr"""
+class_name: Datasource
+module_name: great_expectations.datasource
+execution_engine:
+  module_name: great_expectations.execution_engine
+  class_name: PandasExecutionEngine
+data_connectors:
+  {data_connector_name}:
+    class_name: ConfiguredAssetFilesystemDataConnector
+    assets:
+      {asset_name}:
+        module_name: great_expectations.datasource.data_connector.asset
+        group_names:
+          - batch_num
+          - total_batches
+        pattern: csv_batch_(\d.+)_of_(\d.+)\.csv
+        reader_options:
+          delimiter: ","
+        class_name: Asset
+        base_directory: $data_fixtures_root
+        glob_directive: "*.csv"
+    base_directory: $base_directory
+    module_name: great_expectations.datasource.data_connector
+        """
+
+    context.add_datasource(name=datasource_name, **yaml.load(datasource_config))
+
+    assert context.list_datasources() == [
+        {
+            "class_name": "Datasource",
+            "data_connectors": {
+                data_connector_name: {
+                    "assets": {
+                        asset_name: {
+                            "base_directory": data_relative_path,
+                            "class_name": "Asset",
+                            "glob_directive": "*.csv",
+                            "group_names": ["batch_num", "total_batches"],
+                            "module_name": "great_expectations.datasource.data_connector.asset",
+                            "pattern": "csv_batch_(\\d.+)_of_(\\d.+)\\.csv",
+                        }
+                    },
+                    "base_directory": data_connector_base_directory,
+                    "class_name": "ConfiguredAssetFilesystemDataConnector",
+                    "module_name": "great_expectations.datasource.data_connector",
+                }
+            },
+            "execution_engine": {
+                "class_name": "PandasExecutionEngine",
+                "module_name": "great_expectations.execution_engine",
+            },
+            "module_name": "great_expectations.datasource",
+            "name": "generic_csv_generator",
+        }
+    ]
+    return context
 
 
 @pytest.fixture
@@ -333,7 +467,8 @@ def bobby_columnar_table_multi_batch(empty_data_context):
 
         1. monitor the average number of rows in the tables
 
-        2. use it to validate min/max boundaries of all columns are of the acconting/financial/account type and set up alerts for when things change
+        2. use it to validate min/max boundaries of all columns are of the acconting/financial/account type and set up
+           alerts for when things change
 
         3. have a place to add his domain knowledge of the data (that can also be validated against new data)
 
@@ -1512,6 +1647,7 @@ def rule_without_variables_without_parameters():
     return rule
 
 
+# noinspection PyPep8Naming
 @pytest.fixture
 def rule_with_variables_with_parameters(
     column_Age_domain,
